move methods from interface to idmapping api - new api requires a different set of methods
